+++ 4.py
@@ -1,0 +1,80 @@
+"""
+auth:wb
+date:2019-3-26
+py_version:3.6.4
+4. 对下面的 json 字符串 serial 相同的进行去重。
+
+```javascript
+  [{
+    "name": "张三",
+    "serial": "0001"
+  }, {
+    "name": "李四",
+    "serial": "0002"
+  }, {
+    "name": "王五",
+    "serial": "0003"
+  }, {
+    "name": "王五2",
+    "serial": "0003"
+  }, {
+    "name": "赵四",
+    "serial": "0004"
+  }, {
+    "name": "小明",
+    "serial": "005"
+  }, {
+    "name": "小张",
+    "serial": "006"
+  }, {
+    "name": "小李",
+    "serial": "006"
+  }, {
+    "name": "小李2",
+    "serial": "006"
+  }, {
+    "name": "赵四2",
+    "serial": "0004"
+  }];
+```
+"""
+
+data = [{
+    "name": "张三",
+    "serial": "0001"
+}, {
+    "name": "李四",
+    "serial": "0002"
+}, {
+    "name": "王五",
+    "serial": "0003"
+}, {
+    "name": "王五2",
+    "serial": "0003"
+}, {
+    "name": "赵四",
+    "serial": "0004"
+}, {
+    "name": "小明",
+    "serial": "005"
+}, {
+    "name": "小张",
+    "serial": "006"
+}, {
+    "name": "小李",
+    "serial": "006"
+}, {
+    "name": "小李2",
+    "serial": "006"
+}, {
+    "name": "赵四2",
+    "serial": "0004"
+}]
+rub = set()
+r_data = []
+for i in data:
+    ser = i.get("serial")
+    if ser not in rub:
+        rub.add(ser)
+        r_data.append(i)
+print(r_data)
